// src/types/article.ts
export interface Source {
    id: string;
    name: string;
};
  
export interface Article {
    source: Source;
    author: string;
    title: string;
    description: string;
    url: string;
    urlToImage: string;
    publishedAt: string;
    content: string;
    isPinned?: boolean;
};

export interface NewsResponse {
    status: string;
    totalResults: number;
    articles: Article[];
  };